refactor: pending_activities module
