+++ packages/cryptoplease/lib/features/pending_activities/src/components/payment_request_tile.dart
@@ -2,15 +2,15 @@ import 'package:auto_route/auto_route.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
-import '../../../app/routes.dart';
-import '../../../core/presentation/format_amount.dart';
-import '../../../core/presentation/format_date.dart';
-import '../../../di.dart';
-import '../../../gen/assets.gen.dart';
-import '../../../l10n/device_locale.dart';
-import '../../../l10n/l10n.dart';
-import '../../../ui/colors.dart';
-import '../../payment_request/module.dart';
+import '../../../../app/routes.dart';
+import '../../../../core/presentation/format_amount.dart';
+import '../../../../core/presentation/format_date.dart';
+import '../../../../di.dart';
+import '../../../../gen/assets.gen.dart';
+import '../../../../l10n/device_locale.dart';
+import '../../../../l10n/l10n.dart';
+import '../../../../ui/colors.dart';
+import '../../../payment_request/module.dart';
 import 'styles.dart';
 
 class PaymentRequestTile extends StatefulWidget {
