+++ packages/cryptoplease/lib/features/pending_activities/src/pending_activities_repository.dart
@@ -3,11 +3,11 @@ import 'package:fast_immutable_collections/fast_immutable_collections.dart';
 import 'package:injectable/injectable.dart';
 import 'package:rxdart/rxdart.dart';
 
-import '../../core/tokens/token_list.dart';
-import '../../data/db/db.dart';
-import '../outgoing_direct_payments/module.dart';
-import '../outgoing_split_key_payments/module.dart';
-import '../payment_request/module.dart';
+import '../../../core/tokens/token_list.dart';
+import '../../../data/db/db.dart';
+import '../../outgoing_direct_payments/module.dart';
+import '../../outgoing_split_key_payments/module.dart';
+import '../../payment_request/module.dart';
 import 'pending_activity.dart';
 
 typedef _L = Iterable<PendingActivity>;
